+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * feat: `icp settings autocontainerize true`, always use a docker container for all networks
 * feat: `icp canister migrate-id` - initiate canister ID migration across subnets
 * feat: install proxy canister when starting managed networks with all identities as controllers (or anonymous + default if more than 10 identities)
+  * `icp network status` displays the proxy canister principal
 
 # v0.1.0